tambah modul pembelian, custom modul transaksi, custom modul kelola akun, custom dashboard, etc

+++ app/Http/Controllers/Product/SupplierController.php
@@ -20,18 +20,18 @@ class SupplierController extends Controller
     public function index()
     {
         //Variabel For Title Menu
-        $getTitle = Settings::findOrFail(1);
+        $store_information = Settings::findOrFail(1);
         $supplier = Supplier::all();
 
-        return view('product.supplier.index', compact('getTitle','supplier'));
+        return view('product.supplier.index', compact('store_information','supplier'));
     }
 
     public function create()
     {
         //Variabel For Title Menu
-        $getTitle = Settings::findOrFail(1);
+        $store_information = Settings::findOrFail(1);
 
-        return view('product.supplier.create', compact('getTitle'));
+        return view('product.supplier.create', compact('store_information'));
     }
 
     public function store(Request $request)
@@ -39,12 +39,11 @@ class SupplierController extends Controller
         $rules = [
             'name'         => 'required',
             'address'      => 'nullable',
-            'number_phone' => 'required|unique:suppliers,number_phone',
+            'number_phone' => 'nullable|unique:suppliers,number_phone',
             'desc'         => 'nullable'
         ];
 
         $eMessage = [
-            'name.required'         => 'Nama supplier harus di isi !',
             'number_phone.required' => 'No tlp harus di isi !',
             'number_phone.unique'   => 'No tlp sudah terdata, coba dengan nomor lain !'
         ];
@@ -68,17 +67,17 @@ class SupplierController extends Controller
     public function edit(Supplier $supplier)
     {
         //Variabel For Title Menu
-        $getTitle = Settings::findOrFail(1);
+        $store_information = Settings::findOrFail(1);
 
-        return view('product.supplier.edit', compact('getTitle','supplier'));
+        return view('product.supplier.edit', compact('store_information','supplier'));
     }
 
     public function update(Request $request, Supplier $supplier)
     {
         if ($request->number_phone != $supplier->number_phone){
-            $number_rules = ['required', 'unique:suppliers,number_phone'];
+            $number_rules = ['nullable', 'unique:suppliers,number_phone'];
         }else{
-            $number_rules = ['required'];
+            $number_rules = ['nullable'];
         }
 
         $rules = [
@@ -90,7 +89,6 @@ class SupplierController extends Controller
 
         $eMessage = [
             'name.required'         => 'Nama supplier harus di isi !',
-            'number_phone.required' => 'No tlp harus di isi !',
             'number_phone.unique'   => 'No tlp sudah terdata, coba dengan nomor lain !'
         ];
         
@@ -112,6 +110,12 @@ class SupplierController extends Controller
 
     public function destroy(Supplier $supplier)
     {
+        $count = $supplier->purchaseOrder->count();
+
+        if ($count > 0){
+            return redirect()->back()->with('info', "Tidak dapat menghapus Supplier, terdapat {$count} Supplier di dalam data pembelian produk !");
+        }
+        
         $supplier->delete();
 
         return redirect()->route('supplier.index')->with('success', 'Supplier berhasil dihapus !');
